truthseeker-voice/requirements.txt
# Truthseeker LIVE Dependencies
py-cord[voice]>=2.4.0
opuslib
websockets
python-dotenv
numpy
